requirements.txt
fastapi
uvicorn
rio-tiler
requests
numpy
pillow
paramiko
psutil
sse-starlette
httpx
python-dotenv
psycopg2-binary
asyncpg
sqlalchemy[asyncio]